requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=hd590300_5
ca-certificates=2023.11.17=hbcca054_0
colorama=0.4.6=pypi_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
customtkinter=5.2.1=pypi_0
cycler=0.12.1=pypi_0
darkdetect=0.8.0=pypi_0
debugpy=1.8.0=py311hb755f60_1
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
fonttools=4.47.2=pypi_0
icecream=2.1.3=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
ipykernel=6.28.0=pyhd33586a_0
ipython=8.19.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.0=py311h38be061_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_3
libgomp=13.2.0=h807b86a_3
libnsl=2.0.1=hd590300_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.44.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_3
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
ncurses=6.4=h59595ed_2
nest-asyncio=1.5.8=pyhd8ed1ab_0
numpy=1.26.3=pypi_0
openssl=3.2.0=hd590300_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.2=pyhd8ed1ab_0
platformdirs=4.1.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.7=py311h459d7ec_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.2=pypi_0
pygame-ce=2.4.0=pypi_0
pygame-gui=0.6.9=pypi_0
pygame-menu=4.4.3=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
pyperclip=1.8.2=pypi_0
python=3.11.7=hab00c5b_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-i18n=0.3.9=pypi_0
python_abi=3.11=4_cp311
pytz=2023.3.post1=pypi_0
pyzmq=25.1.2=py311h34ded2d_0
readline=8.2=h8228510_1
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
tornado=6.3.3=py311h459d7ec_1
traitlets=5.14.1=pyhd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
wcwidth=0.2.12=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
zeromq=4.3.5=h59595ed_0
zipp=3.17.0=pyhd8ed1ab_0
